Remove obsolete GitHub checks
* The AMS configuration file was generated from
the YAML file because the original idea was to use
the same YAML file to generate the table in the
docs and in the help texts. But this was never
implemented. We have plans of auto-generating
this content from swagger and hence this method
of generating the configuration file and by
extension this check is no longer needed.

* The docs link checker was added to check if all
links except those in release notes point to
discourse links so that readers using discourse
posts instead of documentation pages will be
redirected to the corresponding discourse posts.
This check is also no longer needed with the move
to RTD.

* The spellcheck and inclusive language check
are part of the automated doc checks that are
included in the starter pack. So there is no need
to have a separate workflow for each check.