feat(many): update for changes coming with appliance v2
The new appliance v2 mainly stays identical as v1 but a few changes
exist. This updates existing content to adapt to the new changes and
removes no longer needed parts.

Note that this drops the stream client tutorial as this is no longer
applicable as is and needs some more rework.