Remove `ams-configuration.md` from the source
This file is automatically generated when the
documentation is built either locally or on RTD.

To keep the source accurate and single-sourced,
this file should be removed from the repository
so that it is only generated from the builds.

The file generated from local builds will be
ignored as it is added to `.gitignore` and
only the file from RTD builds will be published.