Fix the infinite loop
When running `make run`, except for the very
first time, an infinite loop runs for every doc
change because of the auto-generated
`ams-configuration.md` file. The script that
generates this file writes the output irrespective
of whether there are changes to that particular
file or not. This file being generated irrespective
of changes triggers sphinx autobuild which
builds again and again causing an infinite loop.

This commit fixes the infinite loop by writing
to the `ams-configuration.md` file only if there
are changes to it.