fix(release-notes): add a workaround to enable streaming in an insecure way
This change adds a workaround to the bug https://bugs.launchpad.net/anbox-cloud/+bug/2110194
and enable streaming in an insecure way.

Note: we can't not simply replace the incorrect certificate with
the ca certificate by copying it from the anbox-stream-gateway unit
to anbox-cloud-dashboard due to the fact the private IP was not
included into the SAN list when generating the private certificates
for stream-gateway used in the tls verification.
See the following commits for details:
   https://github.com/canonical/anbox-cloud-charms/commit/65cc339f89ef57a7a3f5322865e113cbc8f29f73

Hence so far, the easiest way to unblock people from streaming is
to disable TLS verification on the dashboard side.