fix(upgrade): explicitly specify the channel `1.31/stable` for the easyrsa charm on upgrade
Until now, we specify the revision and channel both in the bundles
```
etcd-ca:
  charm: easyrsa
  revision: 66
  channel: stable
  ...
```
However, this potentially introduced a breakage because Juju fetches
the latest resource revision from the channel rather than the compatible
resource revision corresponding to the charm's pinned revision.
Not specifying the channel in the juju refresh command risks breaking
the etcd-ca and internal-ca charms during upgrade.

Therefore, explicitly specify the 1.31/stable channel for the easyrsa
charm during upgrade.