feat(howto): add permission example for system app howto
Add a permission list example for the how to related to installed system apps.
I needed to use aam and follow this howto recently, and I wasn't sure what the
format for permissions was. This clarifies it greatly by introducing a simple
example based on one of our test apps.